fix: Move auth.ts to src directory and update import paths
- Move auth.ts from root to src/ directory to match @/ path mapping
- Update API route handler import path
- Fix middleware import to use relative path from root
- Resolves 'Module not found: Can't resolve @/auth' error
