+++ web/middleware.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server'
 import type { NextRequest } from 'next/server'
-import { auth } from '@/auth'
+import { auth } from './src/auth'
 
 // Define route categories for protection
 const publicRoutes = [
